requirements.txt
attrs==21.2.0
click==7.1.2
coverage==5.5
GH-IO-stubs==7.7.21140
GH-Util-stubs==7.7.21140
Grasshopper-stubs==7.7.21140
honeybee-core==1.44.18
honeybee-energy==1.75.9
honeybee-radiance==1.41.4
honeybee-radiance-command==1.11.4
honeybee-radiance-folder==2.5.4
honeybee-schema==1.43.6
honeybee-standards==2.0.2
iniconfig==1.1.1
ladybug-core==0.39.4
ladybug-geometry==1.23.7
ladybug-geometry-polyskel==1.3.27
packaging==20.9
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pydantic-openapi-helper==0.2.7
pyparsing==2.4.7
pytest==6.2.4
Rhino==0.0.5
Rhino-stubs==7.7.21140
toml==0.10.2
typing-extensions==3.10.0.0
